configure spotinst to revert on ondemand instances to spot instance at midnight each day
We need this option in order for spotinst to revert any extra on demand instances we have back to spot instances.
This could happen if spotinst was unable to find a spot instance at some point and used ondemand for a hostclass.
It would then never switch back to spot instances even when they bacame available because we were missing this option